\name{AnnualAggLossDevModelInput-class}
\alias{AnnualAggLossDevModelInput-class}
\title{The base class for all aggregate models.}
\description{The base class for all aggregate models.}
\details{\code{AnnualAggLossDevModelInput} is the base class for all aggregate model input objects.  It is derived from \code{LossDevModelInput}.}
\docType{class}
\seealso{\code{\linkS4class{LossDevModelInput}}, \code{\linkS4class{StandardAnnualAggLossDevModelInput}}}
