\name{AnnualAggLossDevModelOutputWithZeros-class}
\alias{AnnualAggLossDevModelOutputWithZeros-class}
\title{The parent of StandardAnnualAggLossDevModelOutputWithZeros and BreakAnnualAggLossDevModelOutputWithZeros.}
\description{The parent of \code{StandardAnnualAggLossDevModelOutputWithZeros} and \code{BreakAnnualAggLossDevModelOutputWithZeros}.}
\details{To avoid creating multiple inheritance (directly), this class is created using \code{setClassUnion}.
The union consists of classes \code{StandardAnnualAggLossDevModelOutputWithZeros} and \code{StandardAnnualAggLossDevModelOutputWithZeros}.}
\docType{class}
\seealso{\code{\linkS4class{LossDevModelOutput}}
\code{\linkS4class{BreakAnnualAggLossDevModelOutputWithZeros}}
\code{\linkS4class{StandardAnnualAggLossDevModelOutputWithZeros}}}
