\name{BreakAnnualAggLossDevModelInput-class}
\alias{BreakAnnualAggLossDevModelInput-class}
\title{The final input class for models with a break.}
\description{The final input class for models with a break.}
\details{\code{BreakAnnualAggLossDevModelInput} is the final input class for models with a break.}
\docType{class}
\seealso{\code{\linkS4class{LossDevModelInput}}
\code{\linkS4class{StandardAnnualAggLossDevModelInput}}}
