\name{calendarYearEffectErrorTracePlot,AnnualAggLossDevModelOutput-method}
\alias{calendarYearEffectErrorTracePlot,AnnualAggLossDevModelOutput-method}
\title{A method to generate the trace plots for select calendar year effect errors.}
\description{A method to generate the trace plots for select calendar year effect errors.}
\value{NULL invisibly.  Only called for the side effect of plotting.}
\docType{methods}
\seealso{\code{\link{calendarYearEffectErrorTracePlot}}
\code{\link{calendarYearEffectErrors}}}
\arguments{\item{object}{The object of type \code{AnnualAggLossDevModelOutput} from which to generate the trace plots.}
\item{elements}{A numeric vector indicating the elements for which to plot the trace.  Valid values are 2 through the total number of exposure years. If NULL, values are selected automatically.}}
