\name{CPI}
\alias{CPI}
\title{Consumer price index (CPI-U) for the United States.}
\usage{CPI}
\description{Consumer price index (CPI-U) for the United States.}
\references{\emph{bls.gov} \cr
\tabular{ll}{
\bold{Not Seasonally Adjusted} \tab \cr
\bold{Series Id:}              \tab CUUR0000SA0 \cr
\bold{Area:}                   \tab U.S. city average \cr
\bold{Item:}                   \tab All items \cr
\bold{Base Period:}            \tab 1982-84=100 \cr
}}
\docType{data}
