\name{dot-onLoad-lossDev}
\title{Intialize the Namespace.}
\usage{.onLoad(libname, pkgname)}
\description{Intialize the Namespace.  Intended for internal use only.}
\details{Currently only sets correct functions for \code{myPkgName} and \code{myLibPath} and loads the \acronym{JAGS} module.}
\alias{.onLoad}
\seealso{\code{\link{.onLoad}}}
\arguments{\item{libname}{The library where the R package is installed.}
\item{pkgname}{The name of the R package.}}
