\name{exposureGrowthTracePlot,AnnualAggLossDevModelOutput-method}
\alias{exposureGrowthTracePlot,AnnualAggLossDevModelOutput-method}
\title{A method to generate the trace plots for select exposure growth rates.}
\description{A method to generate the trace plots for select exposure growth rates.}
\value{NULL invisibly.  Only called for the side effect of plotting.}
\docType{methods}
\seealso{\code{\link{exposureGrowthTracePlot}}
\code{\link{exposureGrowth}}}
\arguments{\item{object}{The object of type \code{AnnualAggLossDevModelOutput} from which to generate the trace plots.}
\item{elements}{A numeric vector indicating the elements for which to plot the trace. If NULL, values are selected automatically.}}
