\name{finalCumulativeDiff}
\alias{finalCumulativeDiff}
\title{A generic function to plot and/or return the difference between final actual and predicted cumulative payments.}
\description{A generic function to plot and/or return the difference between final actual and predicted cumulative payments.}
\value{Mainly called for the side effect of plotting.}
\docType{genericFunction}
\seealso{\code{\link[=finalCumulativeDiff,AnnualAggLossDevModelOutput-method]{finalCumulativeDiff("AnnualAggLossDevModelOutput")}}}
\arguments{\item{object}{The object from which to plot and/or return the difference.}
\item{plot}{A logical value. If \code{TRUE}, the plot is generated and the statistics are returned; otherwise only the statistics are returned.}}
