\name{firstYearInNewRegime,BreakAnnualAggLossDevModelOutput-method}
\alias{firstYearInNewRegime,BreakAnnualAggLossDevModelOutput-method}
\title{A method to plot and/or return the posterior change point.}
\description{A method to plot and/or return the posterior change point.}
\value{Mainly called for the side effect of plotting.  Also returns a named numeric vector with names corresponding to the first year in the new regime and with values corresponding to the density.  Returned invisibly.}
\docType{methods}
\seealso{\code{\link{firstYearInNewRegime}}
\code{\link{firstYearInNewRegimeTracePlot}}}
\arguments{\item{object}{The object from which to plot and/or return the posterior change point estimate.}
\item{plot}{A logical value. If \code{TRUE}, the plot is generated and the statistics are returned; otherwise only the statistics are returned.}}
