\name{firstYearInNewRegimeTracePlot,BreakAnnualAggLossDevModelOutput-method}
\alias{firstYearInNewRegimeTracePlot,BreakAnnualAggLossDevModelOutput-method}
\title{A method to generate the trace plot for the posterior change point.}
\description{A method to generate the trace plot for the posterior change point.}
\value{Only called for the side effect of plotting.}
\docType{methods}
\seealso{\code{\link{firstYearInNewRegimeTracePlot}}}
\seealso{\code{\link{firstYearInNewRegime}}}
\arguments{\item{object}{The object from which to generate the trace plot for the change point estimate.}}
