\name{get.color}
\alias{get.color}
\title{A function to get color values.}
\usage{get.color(i)}
\description{A function to get color values.  Intended for internal use only.}
\details{Currenly, the main feature is that it never returns yellow.
May, in the future, be expanded.}
\value{A color value suitable for ploting commands.}
\arguments{\item{i}{An integer value of 1 or greater.}}
