\name{getExposureYearLabel}
\alias{getExposureYearLabel}
\title{A function to return a "nice" character string for the exposure year label in charts.}
\usage{getExposureYearLabel(object)}
\description{A function to return a "nice" character string for the exposure year label in charts.}
\value{A character.}
\arguments{\item{object}{An object of type \code{AnnualAggLossDevModelInput}.}}
