\name{getTriDim,AnnualAggLossDevModelInput-method}
\alias{getTriDim,AnnualAggLossDevModelInput-method}
\title{A method to get the dimension of the observed triangle.}
\description{A method to get the dimension of the observed triangle.  Intended for internal use only.}
\value{An integer vector of the dimension of the observed triangle.}
\docType{methods}
\seealso{\code{\link{getTriDim}}}
\arguments{\item{object}{An object of type \code{AnnualLossDevModelInput}.}}
