\name{getTriDim}
\alias{getTriDim}
\title{A generic function to get the dimension of the observed triangle.}
\description{A generic function to get the dimension of the observed triangle.  Intended for internal use only.}
\details{The dimension of a triangle is the number of rows and number of columns.}
\value{An integer vector of the dimension of the observed triangle.}
\arguments{\item{object}{The object from which to get the dim of the observed triangle.}}
