\name{lossDev}
\alias{lossDev}
\title{A package for robust stochastic loss development.}
\description{A package for robust stochastic loss development.}
\details{\pkg{lossDev} makes available a Bayesian time series model of loss development, estimated using MCMC.
Features include skewed Student-t distribution with time-varying scale parameter, expert prior for the calendar year effect, and structural break in the consumption path of services.

Please read the vignette for guidance on usage.  And \dQuote{Robust Loss Development Using MCMC} by Schmid, Frank A. for theory.}
\references{Schmid, Frank A., \dQuote{Robust Loss Development Using MCMC,}, 2009}
\docType{package}
