\name{lossDevelopmentFactors,AnnualAggLossDevModelOutput-method}
\alias{lossDevelopmentFactors,AnnualAggLossDevModelOutput-method}
\title{A method to plot and/or return a table of predicted age-to-age loss development factors (or link ratios).}
\description{A method to plot and/or return a table of predicted age-to-age loss development factors (or link ratios).}
\value{Mainly called for the side effect of plotting, but also returns a numeric matrix of plotted statistics.}
\docType{methods}
\seealso{\code{\link{lossDevelopmentFactors}}
\code{\link{tailFactor}}}
\arguments{\item{object}{The object of type /code{AnnualAggLossDevModelOutput} from which to plot and/or return the factors.}
\item{cex.text}{The \code{cex} value supplied to \code{text}. Adjusts the relative size of text.}
\item{linespace}{Adjusts the spacing between observed and predicted values.}
\item{plot}{A logical value. If \code{TRUE}, the plot is generated and the statistics are returned; otherwise only the statistics are returned.}}
