\name{MCPI}
\alias{MCPI}
\title{Medical Care price index for the United States.}
\usage{MCPI}
\description{Medical Care price index for the United States.}
\references{\emph{bls.gov} \cr
\tabular{ll}{
\bold{Not Seasonally Adjusted} \tab \cr
\bold{Series Id:}              \tab CUUR0000SAM \cr
\bold{Area:}                   \tab U.S. city average \cr
\bold{Item:}                   \tab Medical care \cr
\bold{Base Period:}            \tab 1982-84=100 \cr
}}
\docType{data}
