\name{myPkgName}
\alias{myPkgName}
\title{Current Name of the Package.}
\description{Current Name of the Package. Intended for internal use only.}
\details{Set by \code{.onLoad}.}
\value{The current name of the package including version number if the package was installed as such. (i.e. \samp{lossDev})}
