\name{newNodeOutput}
\alias{newNodeOutput}
\title{A method to construct new object of type NodeOutput.}
\usage{newNodeOutput(mcarray)}
\description{A method to construct new object of type \code{NodeOutput}. Intended for internal use only.}
\details{This method will return a valid NodeOutput object.}
\value{An object of class \code{NodeOutput}.}
\seealso{\code{\linkS4class{NodeOutput}}}
\arguments{\item{mcarray}{An S3 object of type \code{mcarray}.}}
