\name{QQPlot,AnnualAggLossDevModelOutput-method}
\alias{QQPlot,AnnualAggLossDevModelOutput-method}
\title{A method to plot a Q-Q plot for models in the lossDev package.}
\description{A method to plot a Q-Q plot for models in the \pkg{lossDev} package.}
\details{This function plots sorted observed log incremental payments vs sorted predicted log incremental payments.
Credible intervals are also plotted.}
\value{NULL. Called for the side effect of plotting.}
\docType{methods}
\seealso{\code{\link{QQPlot}}
\code{\link{triResi}}}
\arguments{\item{object}{The object of type \code{AnnualAggLossDevModelOuput} from which to plot the values.}}
