\name{QQPlot}
\alias{QQPlot}
\title{A generic function to plot a Q-Q plot for models in the lossDev package.}
\description{A generic function to plot a Q-Q plot for models in the \pkg{lossDev} package.}
\details{This function plots sorted observed log incremental payments vs sorted predicted log incremental payments.
Credible intervals are also plotted.}
\value{Mainly called for the side effect of plotting.}
\seealso{\code{\link[=QQPlot,AnnualAggLossDevModelOutput-method]{QQPlot("AnnualAggLossDevModelOutput")}}
\code{\link{triResi}}}
\arguments{\item{object}{The object from which to plot the values.}}
