\name{runLossDevModel}
\alias{runLossDevModel}
\title{A generic function to run models in lossDev.}
\description{A generic function to run models in \pkg{lossDev}.}
\details{Overriding methods must return a valid output object.}
\value{object of class \code{LossDevModelOutput}.}
\docType{genericFunction}
\seealso{\code{\link[=runLossDevModel,LossDevModelInput-method]{runLossDevModel("LossDevModelInput")}}}
\arguments{\item{object}{The object containing the model to estimate.}}
