\name{StandardAnnualAggLossDevModelInput-class}
\alias{StandardAnnualAggLossDevModelInput-class}
\title{The final input class for models without a break.}
\description{The final input class for models without a break.}
\details{\code{StandardAnnualAggLossDevModelInput} is the final input class for models without a break.}
\docType{class}
\seealso{\code{\linkS4class{LossDevModelInput}}
\code{\linkS4class{StandardAnnualAggLossDevModelInput}}}
