\name{StandardAnnualAggLossDevModelOutputWithZeros-class}
\alias{StandardAnnualAggLossDevModelOutputWithZeros-class}
\title{The class to handle incremental payments of zero.}
\description{The class to handle incremental payments of zero.}
\details{\code{StandardAnnualAggLossDevModelOutputWithZeros} is a special class designed to be merged with aggregate annual model objects.}
\docType{class}
\seealso{\code{\linkS4class{LossDevModelOutput}}}
