\name{StandardAnnualAggLossDevModelOutputWithZeros-class}
\alias{StandardAnnualAggLossDevModelOutputWithZeros-class}
\title{The class to handle incremental payments of zero for the standard model.}
\description{The class to handle incremental payments of zero for the standard model.}
\details{\code{StandardAnnualAggLossDevModelOutputWithZeros} is a special class designed to be merged with aggregate annual model objects.
It adds one extra node \code{prob.of.non.zero.payment} to the list of slots.}
\docType{class}
\seealso{\code{\linkS4class{LossDevModelOutput}}
\code{\linkS4class{AnnualAggLossDevModelOutputWithZeros}}}
