\name{standardDeviationOfExposureGrowth}
\alias{standardDeviationOfExposureGrowth}
\title{A generic function to plot and/or return the posterior of the standard deviation of the exposure growth rate for models in lossDev.}
\description{A generic function to plot and/or return the posterior of the standard deviation of the exposure growth rate for models in \pkg{lossDev}.}
\value{Mainly called for the side effect of plotting.}
\docType{genericFunction}
\seealso{\code{\link[=standardDeviationOfExposureGrowth,AnnualAggLossDevModelOutput-method]{standardDeviationOfExposureGrowth("AnnualAggLossDevModelOutput")}}}
\arguments{\item{object}{The object from which to plot and/or return the standard deviation of the exposure growth rate.}
\item{plotDensity}{A logical value. If \code{TRUE}, then the density is plotted. If \code{plotTrace} is also \code{TRUE}, then two plots are generated.  If they are both \code{FALSE}, then only the statistics are returned.}
\item{plotTrace}{A logical value. If \code{TRUE}, then the trace is plotted. If \code{plotDensity} is also \code{TRUE}, then two plots are generated.  If they are both \code{FALSE}, then only the statistics are returned.}}
