\name{CumulativeAutoBodilyInjuryTriangle}
\alias{CumulativeAutoBodilyInjuryTriangle}
\title{An and example of a cumulative loss triangle.}
\usage{CumulativeAutoBodilyInjuryTriangle}
\description{An and example of a cumulative loss triangle.}
\details{Rows pertain to accident years; and columns pertain to development years.
For this triangle accident years range from 1974 to 1991 and development years range from 1 to 18.}
\references{Hayne, Roger M. (2003), "Measurement of Reserve Variability," \emph{Casualty Actuarial Society Forum}, Fall 2003, pp. 141-172,
http://www.casact.org/pubs/forum/03fforum/03ff141.pdf.}
\docType{data}
