\name{firstYearInNewRegimeTracePlot}
\alias{firstYearInNewRegimeTracePlot}
\title{A generic function to generate the trace plot for the posterior change point.}
\description{A generic function to generate the trace plot for the posterior change point.}
\value{Only called for the side effect of plotting.}
\docType{genericFunction}
\seealso{\code{\link[=firstYearInNewRegimeTracePlot,BreakAnnualAggLossDevModelOutput-method]{firstYearInNewRegimeTracePlot("BreakAnnualAggLossDevModelOutput")}}}
\seealso{\code{\link{firstYearInNewRegimeTracePlot}}}
\arguments{\item{object}{The object from which to generate the trace plot for the change point estimate.}}
