\name{IncrementalGeneralLiablityTriangle}
\alias{IncrementalGeneralLiablityTriangle}
\title{An and example of an incremental incurred loss triangle.}
\usage{IncrementalGeneralLiablityTriangle}
\description{An and example of an incremental incurred loss triangle.}
\details{Rows pertain to accident years; and columns pertain to development years.
For this triangle accident years range from 1981 to 1990 and development years range from 1 to 10.

The triangle contains incurred incrementals of Automatic Facultative business in General Liability (excluding Asbestos & Environmental).}
\references{\emph{Historical Loss Development Study}, 1991 edition, published by the Reinsurance Association of America (RAA),
page 96. Quoted from, Thomas Mack (1995),
"Which Stochastic Model is Underlying the Chain Ladder Method," \emph{Casualty Actuarial Society Forum}, Fall 1995, pp. 229-240,
http://www.casact.org/pubs/forum/95fforum/95ff229.pdf.}
\docType{data}
